src/services/face_detection/requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.0.74
pillow==10.0.0
numpy==1.24.3
